requirements.txt
pytorch
numpy
chess
wget